Align max-body with RFC add tests, move tests
